requirements.txt
mcp==1.0.0
fastmcp==0.2.0

cohere==5.0.0

fastapi>=0.110.0
uvicorn[standard]==0.24.0

pydantic>=2.8.0
pydantic-settings>=2.6.1

httpx>=0.27.0
python-dotenv==1.0.0
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1